+++ checkpoint/orbax/checkpoint/experimental/v1/_src/saving/saving.py
@@ -19,51 +19,18 @@
 import threading
 
 from etils import epath
-import jax
-import numpy as np
 import orbax.checkpoint as ocp
 from orbax.checkpoint.experimental.v1._src.context import context as context_lib
-from orbax.checkpoint.experimental.v1._src.context import options as options_lib
+from orbax.checkpoint.experimental.v1._src.handlers import pytree_handler
 from orbax.checkpoint.experimental.v1._src.path import types as path_types
 from orbax.checkpoint.experimental.v1._src.synchronization import types as async_types
 from orbax.checkpoint.experimental.v1._src.tree import types as tree_types
-
-
-def _make_v0_save_args(
-    pytree: tree_types.PyTree,
-    create_array_storage_options_fn: (
-        options_lib.CreateArrayStorageOptionsFn | None
-    ),
-) -> tree_types.PyTree:
-  """Creates v0-compatible `SaveArgs` from v1 storage options."""
-  if create_array_storage_options_fn is None:
-    save_args = None
-  else:
-    array_storage_options_pytree = jax.tree.map_with_path(
-        create_array_storage_options_fn, pytree
-    )
-    save_args = jax.tree.map(
-        lambda v: ocp.SaveArgs(
-            dtype=np.dtype(v.dtype),
-            chunk_byte_size=v.chunk_byte_size,
-            shard_axes=v.shard_axes,
-        ),
-        array_storage_options_pytree,
-    )
-  return save_args
-
-
-def _get_concurrent_gb(concurrent_bytes: int | None) -> int | None:
-  if concurrent_bytes:
-    return max(int(concurrent_bytes / 1e9), 1)
-  return None
 
 
 def save_pytree(
     directory: path_types.PathLike,
     pytree: tree_types.PyTreeOf[tree_types.LeafType],
     *,
-    # Generic arguments.
     force: bool = False,
     custom_metadata: ocp.tree.JsonType | None = None,
 ):
@@ -88,31 +55,13 @@ def save_pytree(
   context = context_lib.get_context()
 
   handler_registry = ocp.handlers.create_default_handler_registry(
-      pytree=ocp.PyTreeCheckpointHandler(
-          use_ocdbt=context.pytree_options.use_ocdbt,
-          use_zarr3=context.pytree_options.use_zarr3,
-          save_concurrent_gb=_get_concurrent_gb(
-              context.pytree_options.save_concurrent_bytes
-          ),
-      )
+      pytree=pytree_handler.create_v0_handler(context)
   )
   ckptr = ocp.Checkpointer(
       ocp.CompositeCheckpointHandler(handler_registry=handler_registry)
   )
-  save_args = _make_v0_save_args(
-      pytree, context.pytree_options.create_array_storage_options_fn
-  )
   args = ocp.args.Composite(
-      pytree=ocp.args.PyTreeSave(
-          pytree,
-          save_args=save_args,
-          ocdbt_target_data_file_size=(
-              context.pytree_options.ocdbt_target_data_file_size
-          ),
-          enable_pinned_host_transfer=(
-              context.pytree_options.enable_pinned_host_transfer
-          ),
-      )
+      pytree=pytree_handler.create_v0_save_args(context, pytree)
   )
   ckptr.save(directory, args=args, force=force, custom_metadata=custom_metadata)
 
@@ -141,7 +90,6 @@ def save_pytree_async(
     directory: path_types.PathLike,
     pytree: tree_types.PyTreeOf[tree_types.LeafType],
     *,
-    # Generic arguments.
     force: bool = False,
     custom_metadata: ocp.tree.JsonType | None = None,
 ) -> async_types.AsyncResponse[None]:
@@ -175,31 +123,13 @@ def save_pytree_async(
   context = context_lib.get_context()
 
   handler_registry = ocp.handlers.create_default_handler_registry(
-      pytree=ocp.PyTreeCheckpointHandler(
-          use_ocdbt=context.pytree_options.use_ocdbt,
-          use_zarr3=context.pytree_options.use_zarr3,
-          save_concurrent_gb=_get_concurrent_gb(
-              context.pytree_options.save_concurrent_bytes
-          ),
-      )
+      pytree=pytree_handler.create_v0_handler(context)
   )
   ckptr = ocp.AsyncCheckpointer(
       ocp.CompositeCheckpointHandler(handler_registry=handler_registry)
   )
-  save_args = _make_v0_save_args(
-      pytree, context.pytree_options.create_array_storage_options_fn
-  )
   args = ocp.args.Composite(
-      pytree=ocp.args.PyTreeSave(
-          pytree,
-          save_args=save_args,
-          ocdbt_target_data_file_size=(
-              context.pytree_options.ocdbt_target_data_file_size
-          ),
-          enable_pinned_host_transfer=(
-              context.pytree_options.enable_pinned_host_transfer
-          ),
-      )
+      pytree=pytree_handler.create_v0_save_args(context, pytree)
   )
   directory = epath.Path(directory)
   ckptr.save(directory, args=args, force=force, custom_metadata=custom_metadata)